MYCLPROJ 3/Photo Automation/requirements.txt
watchdog>=4.0.0
python-pptx>=0.6.23
Pillow>=10.0.0
